requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
coverage==7.6.2
cycler==0.12.1
dj-database-url==0.5.0
Django==5.1.7
django-extensions==2.2.8
django-localflavor==3.0
django-widget-tweaks==1.4.8
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.56.0
google-api-core==2.21.0
google-api-python-client==2.83.0
google-auth==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
jdcal==1.4.1
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.0.3
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.3
python-stdnum==1.20
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
XlsxWriter==1.2.8
